Misc. fixes for testing deployment.
The main fix was that CSSLint was not catching a missing brace on the media query at the bottom of slider.css.
This was causing several critical CSS rules to not show up on non-retina screens. During the frustration, there
may have been other now undocumented modifications to this project.